refactor: move model of AppState to mod.rs
